combine motion example scripts
